requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
pillow==11.3.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
setuptools==80.9.0
watchdog==6.0.0
